combine beams
