+++ startup/smibase/beam.py
@@ -10,7 +10,6 @@ import bluesky.plan_stubs as bps
 from .pilatus import pil1m_pos
 from .beamstop import saxs_bs
 
-from smiclasses.beam  import SMI_SAXS_detector, SMI_WAXS_detector
 from smibase.base import RE
 from smibase.attenuators import (
             att1_1,
@@ -43,6 +42,27 @@ from smibase.pilatus import waxs, pil1M
 from smibase.beamstop import saxs_bs
 from smibase.crls import crl
 from smibase.waxschamber import get_chamber_pressure, chamber_pressure
+
+from ophyd import Signal, Component, Device
+
+# devices to store metadata about detectors
+class SMI_WAXS_detector(Device):
+    prefix = "detector_waxs_"
+    pixel_size = Component(Signal, value=0.172, name=prefix + "pixel_size", kind="hinted")
+    x0_pix = Component(Signal, value=97, name=prefix + "x0_pix", kind="hinted")
+    y0_pix = Component(Signal, value=1386, name=prefix + "y0_pix", kind="hinted")
+    sdd = Component(Signal, value=274.9, name=prefix + "sdd", kind="hinted")
+
+
+class SMI_SAXS_detector(Device):
+    prefix = "detector_saxs_"
+    pixel_size = Component(Signal, value=0.172, name=prefix + "pixel_size", kind="hinted")
+    bs_kind = Component(Signal, value="rod", name=prefix + "bs_kind", kind="hinted")
+    xbs_mask = Component(Signal, value=0, name=prefix + "xbs_mask", kind="hinted")
+    ybs_mask = Component(Signal, value=0, name=prefix + "ybs_mask", kind="hinted")
+    x0_pix = Component(Signal, value=0, name=prefix + "y0_pix", kind="hinted")
+    y0_pix = Component(Signal, value=0, name=prefix + "y0_pix", kind="hinted")
+    sdd = Component(Signal, value=8300, name=prefix + "sdd", kind="hinted")
